feat: implement createAppLoadContext in utils for better session management and modularization across the application
