+++ server.ts
@@ -3,7 +3,7 @@
 import * as remixBuild from "virtual:remix/server-build";
 import { storefrontRedirect } from "@shopify/hydrogen";
 import { createRequestHandler } from "@shopify/remix-oxygen";
-import { createAppLoadContext } from "~/lib/context";
+import { createAppLoadContext } from "~/lib/utils/context";
 
 /**
  * Export a fetch handler in module format.
